feat(app): render activity table for public facing routes
