+++ app/src/components/ActivityColumnFilters.tsx
@@ -10,7 +10,7 @@ import type {
   ActivitySort,
   ActivityStatusFilter,
   ActivityTypeFilter,
-} from "@/lib/profile";
+} from "@bluelearn/schemas";
 import { cn } from "@/lib/utils";
 import { Input } from "@/components/ui/input";
 import { Checkbox } from "@/components/ui/checkbox";
