+++ app/src/lib/api/identity.ts
@@ -17,21 +17,7 @@ export async function deleteMyAccount() {
   await assertOk(res);
 }
 
-export async function getMyStats({ signal }: FetchOptions = {}) {
-  const res = await me.stats.$get(undefined, { init: { signal } });
-  await assertOk(res);
-
-  return await res.json();
-}
-
-export async function getMyActivity({ signal }: FetchOptions = {}) {
-  const res = await me.activity.$get(undefined, { init: { signal } });
-  await assertOk(res);
-
-  return await res.json();
-}
-
-export async function getPublicProfile(
+export async function getProfilePage(
   username: string,
   { signal }: FetchOptions = {}
 ) {
